refactor: reorganize pages evaluation folders for semantics
